refactor: layout.tsx에서 프로필 드롭다운 컴포넌트 가져오기 업데이트
